refactor(run): refactor for new lint rules
